Reuse Button inside BackButton component
