--- website/frontend/src/common/components/BackButton/BackButton.module.scss
@@ -1,10 +1,0 @@
-.button {
-  padding: 0;
-  display: flex;
-
-  &::after {
-    content: url('/assets/images/arrow_back.svg');
-    // to get rid of extra space
-    font-size: 0;
-  }
-}
