+++ website/frontend/src/common/components/Button/Button.tsx
@@ -1,23 +1,7 @@
-import React, { ButtonHTMLAttributes } from 'react';
+import React from 'react';
 import clsx from 'clsx';
+import { Props } from './types';
 import styles from './Button.module.scss';
-
-interface BaseProps extends ButtonHTMLAttributes<HTMLButtonElement> {
-  text?: string;
-  icon?: string;
-  color?: 'success' | 'error' | 'main';
-  size?: 'normal' | 'small';
-}
-
-interface TextProps extends BaseProps {
-  text: string;
-}
-
-interface IconProps extends BaseProps {
-  icon: string;
-}
-
-type Props = TextProps | IconProps;
 
 const Button = ({ text, icon, className, color, size, ...attrs }: Props) => {
   const colorName = color || (text ? 'success' : '');
